Basic workspace setup done
I've made a template workspace with the necessary things added in for
the catkin_ws, so if anyone needs to use a blank catkin_ws copy over
that folder into your system and fiddle around.

I've also added the e-Yantra resource for how to set the workspace up
for reference.